+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#57](https://github.com/jcpsantiago/thearqivist/issues/57) Switch to clojure.java-time
 * [#47](https://github.com/jcpsantiago/thearqivist/issues/47) Save channel `once`
 * Add `app.json` and define healthchecks for deployments with Dokku/Heroku
+* Correctly set the logging as pretty EDN for dev, and JSON for prod
 
 ## 0.1.0 - 2023-04-20